Switch backends package to registry pattern
