+++ pkg/backends/options.go
@@ -1,19 +1,20 @@
-package internal
+package backends
 
 import (
-	clientSetV "github.com/hellofresh/kangal/pkg/kubernetes/generated/clientset/versioned"
 	"go.uber.org/zap"
 
 	"k8s.io/client-go/kubernetes"
 	coreListersV1 "k8s.io/client-go/listers/core/v1"
+
+	clientSetV "github.com/hellofresh/kangal/pkg/kubernetes/generated/clientset/versioned"
 )
 
 // Option allows to configure backends
-type Option func(*Registry)
+type Option func(*registry)
 
 // WithLogger adds given logger to each registered backend that implements BackendLogger
 func WithLogger(logger *zap.Logger) Option {
-	return func(b *Registry) {
+	return func(b *registry) {
 		for _, item := range b.registry {
 			if iface, ok := item.(BackendSetLogger); ok {
 				iface.SetLogger(logger)
@@ -24,7 +25,7 @@ func WithLogger(logger *zap.Logger) Option {
 
 // WithPodAnnotations adds given logger to each registered backend that implements BackendLogger
 func WithPodAnnotations(podAnnotations map[string]string) Option {
-	return func(b *Registry) {
+	return func(b *registry) {
 		for _, item := range b.registry {
 			if iface, ok := item.(BackendSetPodAnnotations); ok {
 				iface.SetPodAnnotations(podAnnotations)
@@ -35,7 +36,7 @@ func WithPodAnnotations(podAnnotations map[string]string) Option {
 
 // WithKubeClientSet adds given kubeClientSet to each registered backend that implements BackendKubeClientSet
 func WithKubeClientSet(kubeClientSet kubernetes.Interface) Option {
-	return func(b *Registry) {
+	return func(b *registry) {
 		for _, item := range b.registry {
 			if iface, ok := item.(BackendSetKubeClientSet); ok {
 				iface.SetKubeClientSet(kubeClientSet)
@@ -46,7 +47,7 @@ func WithKubeClientSet(kubeClientSet kubernetes.Interface) Option {
 
 // WithKangalClientSet adds given kangalClientSet to each registered backend that implements BackendKangalClientSet
 func WithKangalClientSet(kangalClientSet clientSetV.Interface) Option {
-	return func(b *Registry) {
+	return func(b *registry) {
 		for _, item := range b.registry {
 			if iface, ok := item.(BackendSetKangalClientSet); ok {
 				iface.SetKangalClientSet(kangalClientSet)
@@ -57,7 +58,7 @@ func WithKangalClientSet(kangalClientSet clientSetV.Interface) Option {
 
 // WithNamespaceLister adds given namespaceLister to each registered backend that implements BackendNamespaceLister
 func WithNamespaceLister(namespaceLister coreListersV1.NamespaceLister) Option {
-	return func(b *Registry) {
+	return func(b *registry) {
 		for _, item := range b.registry {
 			if iface, ok := item.(BackendSetNamespaceLister); ok {
 				iface.SetNamespaceLister(namespaceLister)
