+++ pkg/backends/registry_test.go
@@ -1,4 +1,4 @@
-package internal
+package backends
 
 import (
 	"testing"
